Separate create community tabs into pages.
